requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-colorfield==0.13.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
google==3.0.0
google-api-core==2.24.2
google-auth==2.39.0
google-cloud==0.34.0
google-cloud-core==2.4.3
google-cloud-secret-manager==2.23.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jiter==0.9.0
msgpack==1.1.0
openai==1.75.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.9.0
pyOpenSSL==25.0.0
pytz==2025.2
redis==5.2.1
requests==2.32.3
rsa==4.9.1
service-identity==24.2.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
zope.interface==7.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.13.4
billiard==4.2.1
cachetools==5.5.2
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-colorfield==0.13.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django_neomodel==0.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
google==3.0.0
google-api-core==2.24.2
google-auth==2.39.0
google-cloud==0.34.0
google-cloud-core==2.4.3
google-cloud-secret-manager==2.23.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jiter==0.9.0
kombu==5.5.3
msgpack==1.1.0
neo4j==5.19.0
neomodel==5.3.3
openai==1.75.0
pillow==11.2.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==5.2.1
requests==2.32.3
rsa==4.9.1
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
